calibration of warmup using NM/PSO works

+++ src/pySODM/models/validation.py
@@ -10,7 +10,7 @@ def date_to_diff(actual_start_date, end_date):
     Convert date string to int (i.e. number of days since day 0 of simulation,
     which is warmup days before actual_start_date)
     """
-    return int((pd.to_datetime(end_date)-pd.to_datetime(actual_start_date))/pd.to_timedelta('1D'))
+    return float((pd.to_datetime(end_date)-pd.to_datetime(actual_start_date))/pd.to_timedelta('1D'))
 
 def validate_simulation_time(time, warmup):
     """Validates the simulation time of the sim() function. """
